+++ src/lib.rs
@@ -15,7 +15,7 @@ use crate::{
     collection::MenuItemCollection,
     interaction::{
         programmed::Programmed, Action, InputAdapter, InputAdapterSource, InputResult, InputState,
-        Interaction,
+        Interaction, Navigation,
     },
     selection_indicator::{
         style::{line::Line as LineIndicator, IndicatorStyle},
@@ -40,45 +40,7 @@ use embedded_text::{
     TextBox,
 };
 
-use crate::interaction::Navigation;
 pub use embedded_menu_macros::SelectValue;
-
-/// Marker trait necessary to avoid a "conflicting implementations" error.
-pub trait Marker {}
-
-pub trait MenuListItem<R>: Marker + View {
-    /// Returns the value of the selected item, without interacting with it.
-    fn value_of(&self) -> R;
-
-    fn interact(&mut self) -> R;
-
-    fn set_style<S, IT, P, C>(&mut self, style: &MenuStyle<S, IT, P, R, C>)
-    where
-        S: IndicatorStyle,
-        IT: InputAdapterSource<R>,
-        P: SelectionIndicatorController,
-        C: Theme;
-
-    /// Returns whether the list item is selectable.
-    ///
-    /// If this returns false, the list item will not be interactable and user navigation will skip
-    /// over it.
-    fn selectable(&self) -> bool {
-        true
-    }
-
-    fn draw_styled<S, IT, P, D, C>(
-        &self,
-        style: &MenuStyle<S, IT, P, R, C>,
-        display: &mut D,
-    ) -> Result<(), D::Error>
-    where
-        S: IndicatorStyle,
-        IT: InputAdapterSource<R>,
-        P: SelectionIndicatorController,
-        D: DrawTarget<Color = BinaryColor>,
-        C: Theme;
-}
 
 #[derive(Copy, Clone, Debug)]
 pub enum DisplayScrollbar {